+++ exercises/variables/variables2.rs
@@ -3,7 +3,7 @@
 
 
 fn main() {
-    let x: i32 = 10;
+    let x: i8 = 10;
     if x == 10 {
         println!("x is ten!");
     } else {
